lib-http: Avoid compiler warnings in printf() strings. Cast fields explicitly as long, which should be large enough.

+++ src/lib-http/http-server-connection.c
@@ -806,10 +806,10 @@ http_server_connection_create(struct http_server *server,
 			if (net_getunixcred(fd_in, &cred) < 0) {
 				name = t_strdup_printf("[%u]", id);
 			} else if (cred.pid == (pid_t)-1) {
-				name = t_strdup_printf("unix:uid=%u [%u]", cred.uid, id);
+				name = t_strdup_printf("unix:uid=%ld [%u]", (long)cred.uid, id);
 			} else {
 				name = t_strdup_printf
-					("unix:pid=%u,uid=%u [%u]", cred.pid, cred.uid, id);
+					("unix:pid=%ld,uid=%ld [%u]", (long)cred.pid, (long)cred.uid, id);
 			}
 		} else if (addr.family == AF_INET6) {
 			name = t_strdup_printf("[%s]:%u [%u]", net_ip2addr(&addr), port, id);
